Merge

+++ src/main/java/org/spongepowered/mod/effect/particle/SpongeParticleEffect.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of Sponge, licensed under the MIT License (MIT).
  *
- * Copyright (c) SpongePowered.org <http://www.spongepowered.org>
+ * Copyright (c) SpongePowered <https://www.spongepowered.org>
  * Copyright (c) contributors
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -24,22 +24,22 @@
  */
 package org.spongepowered.mod.effect.particle;
 
-import java.awt.Color;
-
+import com.flowpowered.math.vector.Vector3d;
 import org.spongepowered.api.effect.particle.ParticleEffect;
 import org.spongepowered.api.item.inventory.ItemStack;
 
-import com.flowpowered.math.vector.Vector3f;
+import java.awt.Color;
 
 public class SpongeParticleEffect implements ParticleEffect {
+
     private SpongeParticleType type;
 
-    private Vector3f motion;
-    private Vector3f offset;
+    private Vector3d motion;
+    private Vector3d offset;
 
     private int count;
 
-    public SpongeParticleEffect(SpongeParticleType type, Vector3f motion, Vector3f offset, int count) {
+    public SpongeParticleEffect(SpongeParticleType type, Vector3d motion, Vector3d offset, int count) {
         this.motion = motion;
         this.offset = offset;
         this.count = count;
@@ -52,12 +52,12 @@ public class SpongeParticleEffect implements ParticleEffect {
     }
 
     @Override
-    public Vector3f getMotion() {
+    public Vector3d getMotion() {
         return this.motion;
     }
 
     @Override
-    public Vector3f getOffset() {
+    public Vector3d getOffset() {
         return this.offset;
     }
 
@@ -67,9 +67,10 @@ public class SpongeParticleEffect implements ParticleEffect {
     }
 
     public static class Colored extends SpongeParticleEffect implements ParticleEffect.Colorable {
+
         private Color color;
 
-        public Colored(SpongeParticleType type, Vector3f motion, Vector3f offset, Color color, int count) {
+        public Colored(SpongeParticleType type, Vector3d motion, Vector3d offset, Color color, int count) {
             super(type, motion, offset, count);
             this.color = color;
         }
@@ -82,9 +83,10 @@ public class SpongeParticleEffect implements ParticleEffect {
     }
 
     public static class Resized extends SpongeParticleEffect implements ParticleEffect.Resizable {
+
         private float size;
 
-        public Resized(SpongeParticleType type, Vector3f motion, Vector3f offset, float size, int count) {
+        public Resized(SpongeParticleType type, Vector3d motion, Vector3d offset, float size, int count) {
             super(type, motion, offset, count);
             this.size = size;
         }
@@ -97,9 +99,10 @@ public class SpongeParticleEffect implements ParticleEffect {
     }
 
     public static class Note extends SpongeParticleEffect implements ParticleEffect.Note {
+
         private float note;
 
-        public Note(SpongeParticleType type, Vector3f motion, Vector3f offset, float note, int count) {
+        public Note(SpongeParticleType type, Vector3d motion, Vector3d offset, float note, int count) {
             super(type, motion, offset, count);
             this.note = note;
         }
@@ -112,9 +115,10 @@ public class SpongeParticleEffect implements ParticleEffect {
     }
 
     public static class Materialized extends SpongeParticleEffect implements ParticleEffect.Material {
+
         private ItemStack item;
 
-        public Materialized(SpongeParticleType type, Vector3f motion, Vector3f offset, ItemStack item, int count) {
+        public Materialized(SpongeParticleType type, Vector3d motion, Vector3d offset, ItemStack item, int count) {
             super(type, motion, offset, count);
             this.item = item;
         }
